requirements.txt
agate==1.6.3
anyio==3.6.2
attrs==22.1.0
Authlib==1.1.0
Babel==2.10.3
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==7.1.2
colorama==0.4.5
cryptography==38.0.3
db-dtypes==1.0.1
dbt-bigquery==1.2.0
dbt-core==1.2.0
dbt-extractor==0.4.1
Flask==1.1.4
future==0.18.2
gcloud==0.18.3
google-api-core==2.8.2
google-auth==2.10.0
google-auth-oauthlib==0.5.2
google-cloud-bigquery==2.34.4
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.2
google-cloud-secret-manager==2.10.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.6
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
h11==0.12.0
hologram==0.0.15
html5lib==1.1
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
idna==3.3
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
leather==0.3.4
Logbook==1.5.3
lxml==4.9.1
MarkupSafe==2.0.1
mashumaro==2.9
minimal-snowplow-tracker==0.0.2
msgpack==1.0.4
netsuite==0.9.0
netsuitesdk==2.14.0
networkx==2.8.5
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pandas-gbq==0.17.6
parsedatetime==2.4
platformdirs==2.5.4
proto-plus==1.22.0
protobuf==4.21.12
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pydata-google-auth==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytimeparse==1.1.8
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rsa==4.9
session-info==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stdlib-list==0.8.0
text-unidecode==1.3
typing_extensions==4.3.0
urllib3==1.26.11
webencodings==0.5.1
Werkzeug==1.0.1
xlrd==2.0.1
zeep==4.2.0